@@ -1,25 +1,41 @@
 module github.com/schollz/croc/v9
 
-go 1.13
+go 1.17
 
 require (
 	github.com/OneOfOne/xxhash v1.2.5 // indirect
 	github.com/buger/jsonparser v1.1.1 // Added by Mawoka to Parse JSON coming from GitHub-Api
 	github.com/cespare/xxhash v1.1.0
-	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/denisbrodbeck/machineid v1.0.1
 	github.com/kalafut/imohash v1.0.2
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/schollz/cli/v2 v2.2.1
 	github.com/schollz/logger v1.2.0
 	github.com/schollz/mnemonicode v1.0.1
-	github.com/schollz/pake/v3 v3.0.2
-	github.com/schollz/peerdiscovery v1.6.6
-	github.com/schollz/progressbar/v3 v3.8.2
-	github.com/spaolacci/murmur3 v1.1.0 // indirect
+	github.com/schollz/pake/v3 v3.0.4
+	github.com/schollz/peerdiscovery v1.6.11
+	github.com/schollz/progressbar/v3 v3.8.6
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20210616213533-5ff15b29337e
-	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
+	golang.org/x/crypto v0.0.0-20220321153916-2c7772ba3064
+	golang.org/x/net v0.0.0-20220325170049-de3da57026de
+	golang.org/x/time v0.0.0-20220224211638-0e9765cccd65
+)
+
+require (
+	github.com/OneOfOne/xxhash v1.2.5 // indirect
+	github.com/cpuguy83/go-md2man/v2 v2.0.1 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/pretty v0.1.0 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/spaolacci/murmur3 v1.1.0 // indirect
+	github.com/tscholl2/siec v0.0.0-20210707234609-9bdfc483d499 // indirect
+	github.com/twmb/murmur3 v1.1.6 // indirect
+	golang.org/x/sys v0.0.0-20220330033206-e17cdc41300f // indirect
+	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211 // indirect
+	golang.org/x/text v0.3.7 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
