@@ -1,8 +1,9 @@
 package croc
 
 import (
-	"io/ioutil"
 	"os"
+	"path"
+	"path/filepath"
 	"sync"
 	"testing"
 	"time"
@@ -18,11 +19,11 @@ import (
 func init() {
 	log.SetLevel("trace")
 
-	go tcp.Run("debug", "8081", "pass123", "8082,8083,8084,8085")
-	go tcp.Run("debug", "8082", "pass123")
-	go tcp.Run("debug", "8083", "pass123")
-	go tcp.Run("debug", "8084", "pass123")
-	go tcp.Run("debug", "8085", "pass123")
+	go tcp.Run("debug", "localhost", "8281", "pass123", "8282,8283,8284,8285")
+	go tcp.Run("debug", "localhost", "8282", "pass123")
+	go tcp.Run("debug", "localhost", "8283", "pass123")
+	go tcp.Run("debug", "localhost", "8284", "pass123")
+	go tcp.Run("debug", "localhost", "8285", "pass123")
 	time.Sleep(1 * time.Second)
 }
 
@@ -34,8 +35,8 @@ func TestCrocReadme(t *testing.T) {
 		IsSender:      true,
 		SharedSecret:  "8123-testingthecroc",
 		Debug:         true,
-		RelayAddress:  "localhost:8081",
-		RelayPorts:    []string{"8081"},
+		RelayAddress:  "localhost:8281",
+		RelayPorts:    []string{"8281"},
 		RelayPassword: "pass123",
 		Stdout:        false,
 		NoPrompt:      true,
@@ -52,7 +53,7 @@ func TestCrocReadme(t *testing.T) {
 		IsSender:      false,
 		SharedSecret:  "8123-testingthecroc",
 		Debug:         true,
-		RelayAddress:  "localhost:8081",
+		RelayAddress:  "localhost:8281",
 		RelayPassword: "pass123",
 		Stdout:        false,
 		NoPrompt:      true,
@@ -70,9 +71,11 @@ func TestCrocReadme(t *testing.T) {
 	body, _ := ioutil.ReadAll(http_resp.Body)
 	gh_version, _ := jsonparser.GetString([]byte(fmt.Sprintf("%s\n", body)), "name")
 	go func() {
-		err := sender.Send(TransferOptions{
-			PathToFiles: []string{"../../README.md"},
-		}, gh_version)
+		filesInfo, errGet := GetFilesInfo([]string{"../../README.md"})
+		if errGet != nil {
+			t.Errorf("failed to get minimal info: %v", errGet)
+		}
+		err := sender.Send(filesInfo, gh_version)
 		if err != nil {
 			t.Errorf("send failed: %v", err)
 		}
@@ -88,6 +91,80 @@ func TestCrocReadme(t *testing.T) {
 	}()
 
 	wg.Wait()
+}
+
+func TestCrocSymlink(t *testing.T) {
+	pathName := "../link-in-folder"
+	defer os.RemoveAll(pathName)
+	os.MkdirAll(pathName, 0755)
+	os.Symlink("../../README.md", filepath.Join(pathName, "README.link"))
+
+	log.Debug("setting up sender")
+	sender, err := New(Options{
+		IsSender:      true,
+		SharedSecret:  "8124-testingthecroc",
+		Debug:         true,
+		RelayAddress:  "localhost:8281",
+		RelayPorts:    []string{"8281"},
+		RelayPassword: "pass123",
+		Stdout:        false,
+		NoPrompt:      true,
+		DisableLocal:  true,
+		Curve:         "siec",
+		Overwrite:     true,
+	})
+	if err != nil {
+		panic(err)
+	}
+
+	log.Debug("setting up receiver")
+	receiver, err := New(Options{
+		IsSender:      false,
+		SharedSecret:  "8124-testingthecroc",
+		Debug:         true,
+		RelayAddress:  "localhost:8281",
+		RelayPassword: "pass123",
+		Stdout:        false,
+		NoPrompt:      true,
+		DisableLocal:  true,
+		Curve:         "siec",
+		Overwrite:     true,
+	})
+	if err != nil {
+		panic(err)
+	}
+
+	var wg sync.WaitGroup
+	wg.Add(2)
+	go func() {
+		filesInfo, errGet := GetFilesInfo([]string{pathName})
+		if errGet != nil {
+			t.Errorf("failed to get minimal info: %v", errGet)
+		}
+		err := sender.Send(filesInfo)
+		if err != nil {
+			t.Errorf("send failed: %v", err)
+		}
+		wg.Done()
+	}()
+	time.Sleep(100 * time.Millisecond)
+	go func() {
+		err := receiver.Receive()
+		if err != nil {
+			t.Errorf("receive failed: %v", err)
+		}
+		wg.Done()
+	}()
+
+	wg.Wait()
+
+	s, err := filepath.EvalSymlinks(path.Join(pathName, "README.link"))
+	if s != "../../README.md" {
+		t.Errorf("symlink failed to transfer in folder")
+	}
+	if err != nil {
+		t.Errorf("symlink transfer failed: %s", err.Error())
+	}
 }
 
 func TestCrocLocal(t *testing.T) {
@@ -143,6 +220,12 @@ func TestCrocLocal(t *testing.T) {
 			PathToFiles:      []string{"../../LICENSE", "touched"},
 			KeepPathInRemote: false,
 		}, gh_version)
+
+		filesInfo, errGet := GetFilesInfo([]string{"../../LICENSE", "touched"})
+		if errGet != nil {
+			t.Errorf("failed to get minimal info: %v", errGet)
+		}
+		err := sender.Send(filesInfo, gh_version)
 		if err != nil {
 			t.Errorf("send failed: %v", err)
 		}
@@ -162,7 +245,7 @@ func TestCrocLocal(t *testing.T) {
 
 func TestCrocError(t *testing.T) {
 	content := []byte("temporary file's content")
-	tmpfile, err := ioutil.TempFile("", "example")
+	tmpfile, err := os.CreateTemp("", "example")
 	if err != nil {
 		panic(err)
 	}
@@ -194,10 +277,11 @@ func TestCrocError(t *testing.T) {
 	http_resp, _ := http.Get("https://api.github.com/repos/schollz/croc/releases/latest")
 	body, _ := ioutil.ReadAll(http_resp.Body)
 	gh_version, _ := jsonparser.GetString([]byte(fmt.Sprintf("%s\n", body)), "name")
-	err = sender.Send(TransferOptions{
-		PathToFiles:      []string{tmpfile.Name()},
-		KeepPathInRemote: true,
-	}, gh_version)
+	filesInfo, errGet := GetFilesInfo([]string{tmpfile.Name()})
+	if errGet != nil {
+		t.Errorf("failed to get minimal info: %v", errGet)
+	}
+	err = sender.Send(filesInfo, gh_version)
 	log.Debug(err)
 	assert.NotNil(t, err)
 
