@@ -3,21 +3,21 @@ package utils
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/schollz/croc/v9/src/models"
 	"github.com/stretchr/testify/assert"
 )
+
+const TCP_BUFFER_SIZE = 1024 * 64
 
 var bigFileSize = 75000000
 
 func bigFile() {
-	ioutil.WriteFile("bigfile.test", bytes.Repeat([]byte("z"), bigFileSize), 0666)
+	os.WriteFile("bigfile.test", bytes.Repeat([]byte("z"), bigFileSize), 0666)
 }
 
 func BenchmarkMD5(b *testing.B) {
@@ -63,7 +63,7 @@ func BenchmarkMissingChunks(b *testing.B) {
 	bigFile()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		MissingChunks("bigfile.test", int64(bigFileSize), models.TCP_BUFFER_SIZE/2)
+		MissingChunks("bigfile.test", int64(bigFileSize), TCP_BUFFER_SIZE/2)
 	}
 }
 
@@ -119,7 +119,7 @@ func TestMissingChunks(t *testing.T) {
 	rand.Seed(1)
 	bigBuff := make([]byte, fileSize)
 	rand.Read(bigBuff)
-	ioutil.WriteFile("missing.test", bigBuff, 0644)
+	os.WriteFile("missing.test", bigBuff, 0644)
 	empty := make([]byte, chunkSize)
 	f, err := os.OpenFile("missing.test", os.O_RDWR, 0644)
 	assert.Nil(t, err)
@@ -139,7 +139,7 @@ func TestMissingChunks(t *testing.T) {
 	os.Remove("missing.test")
 
 	content := []byte("temporary file's content")
-	tmpfile, err := ioutil.TempFile("", "example")
+	tmpfile, err := os.CreateTemp("", "example")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -171,7 +171,7 @@ func TestMissingChunks(t *testing.T) {
 
 func TestHashFile(t *testing.T) {
 	content := []byte("temporary file's content")
-	tmpfile, err := ioutil.TempFile("", "example")
+	tmpfile, err := os.CreateTemp("", "example")
 	if err != nil {
 		log.Fatal(err)
 	}
